+++ test/chunking-form/samples/allow-extension-exports/_expected/system/generated-lib2.js
@@ -1,0 +1,13 @@
+System.register(['./generated-lib1.js'], (function (exports) {
+	'use strict';
+	return {
+		setters: [function (module) {
+			exports({ value1: module.v, value2: module.a });
+		}],
+		execute: (function () {
+
+
+
+		})
+	};
+}));