+++ test/chunking-form/samples/circular-entry-points3/_expected/amd/main2.js
@@ -1,8 +1,27 @@
-define(['exports', './generated-main3'], (function (exports, main2) { 'use strict';
+define(['exports'], (function (exports) { 'use strict';
 
+  let C$1 = class C {
+    fn (num) {
+      console.log(num - p);
+    }
+  };
 
+  var p$1 = 43;
 
-	exports.p = main2.p;
-	exports.p2 = main2.p$1;
+  new C$1().fn(p$1);
+
+  class C {
+    fn (num) {
+      console.log(num - p$1);
+    }
+  }
+
+  var p = 42;
+
+  new C().fn(p);
+
+  exports.C = C$1;
+  exports.p = p$1;
+  exports.p2 = p;
 
 }));