+++ test/chunking-form/samples/reexport-from-entry/_expected/es/generated-main.js
@@ -1,6 +1,0 @@
-let a = {};
-let b = {};
-
-let c = {};
-
-export { a, b, c };