+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 # rollup changelog
 
+## 4.39.0
+
+_2025-04-02_
+
+### Features
+
+- Do not create separate facade chunks if a chunk would contain several entry modules that allow export extension if there are no export name conflicts (#5891)
+
+### Bug Fixes
+
+- Mark the `id` property as optional in the filter for the `resolveId` hook (#5896)
+
+### Pull Requests
+
+- [#5891](https://github.com/rollup/rollup/pull/5891): chunk: merge allow-extension modules (@wmertens, @lukastaegert)
+- [#5893](https://github.com/rollup/rollup/pull/5893): chore(deps): update dependency vite to v6.2.4 [security] (@renovate[bot])
+- [#5896](https://github.com/rollup/rollup/pull/5896): fix: resolveId id filter is optional (@sapphi-red)
+
+## 4.38.0
+
+_2025-03-29_
+
+### Features
+
+- Support `.filter` option in `resolveId`, `load` and `transform` hooks (#5882)
+
+### Pull Requests
+
+- [#5882](https://github.com/rollup/rollup/pull/5882): Add support for hook filters (@sapphi-red)
+- [#5894](https://github.com/rollup/rollup/pull/5894): fix(deps): lock file maintenance minor/patch updates (@renovate[bot])
+- [#5895](https://github.com/rollup/rollup/pull/5895): chore(deps): update dependency eslint-plugin-unicorn to v58 (@renovate[bot])
+
 ## 4.37.0
 
 _2025-03-23_